Move cloud CLI tools to common file
